+++ frontend/src/config/axiosConfig.js
@@ -2,5 +2,5 @@
 // For production please use environment variables: REACT_APP_AXIOS_URL
 
 export const AXIOS_CONFIG = {
-    URL: "http://localhost:54000"
+    URL: "http://localhost:54001"
 }